Reworking silent auth based on the auth0-js package
