+++ src/app/helpers/auth.ts
@@ -1,3 +1,6 @@
+import {AuthOptions, WebAuth} from 'auth0-js';
+const appConfig = require('../../../config/main');
+
 export function getToken(): string|null {
   return localStorage.getItem('token');
 }
@@ -92,4 +95,16 @@ export function getBeneficiaryScope(): string|null {
     return null;
   }
   return decoded.app_metadata.scope;
+}
+
+export function getWebAuth(): WebAuth {
+  const options: AuthOptions = {
+    domain: appConfig.app.auth.domain,
+    clientID: appConfig.app.auth.clientID,
+    scope: appConfig.app.auth.scope,
+    responseType: 'token id_token',
+    redirectUri: `${appConfig.app.root}/login`,
+  };
+
+  return new WebAuth(options);
 }
