+++ src/index.ejs
@@ -7,6 +7,5 @@
   </head>
   <body>
       <div id="app"></div>
-      <iframe src="/tokenrefresh.html" id="refresh-iframe" style="display:none;"></iframe>
   </body>
 </html>
